refactor selectors and tests
