+++ src/shared/models/competency-display-name.ts
@@ -2,4 +2,5 @@ export enum CompetencyDisplayName {
   CONTROLLED_STOP = 'Controlled Stop',
   EYESIGHT_TEST = 'Eyesight Test',
   SHOW_ME_TELL_ME = 'Show Me/Tell Me',
+  UNCOUPLE_RECOUPLE = 'Uncouple Recouple',
 }
